toss new stuff into old file
